move send_to_kafka to utils
