+++ corehq/apps/change_feed/pillow.py
@@ -1,14 +1,12 @@
-import json
 from kafka import KeyedProducer
-from kafka.common import KafkaUnavailableError, LeaderNotAvailableError, FailedPayloadsError
-import time
+from kafka.common import KafkaUnavailableError
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.change_feed import data_sources
 from corehq.apps.change_feed.connection import get_kafka_client
 from corehq.apps.change_feed.topics import get_topic
+from corehq.apps.change_feed.utils import send_to_kafka
 from corehq.apps.users.models import CommCareUser
 from corehq.util.couchdb_management import couch_config
-from corehq.util.soft_assert import soft_assert
 from couchforms.models import all_known_formlike_doc_types
 import logging
 from pillowtop.checkpoints.manager import PillowCheckpoint, get_django_checkpoint_store, \
@@ -45,31 +43,7 @@ class KafkaProcessor(PillowProcessor):
                 is_deletion=change.deleted,
             )
 
-            def _send_to_kafka():
-                # closures
-                self._producer.send_messages(
-                    bytes(get_topic(document_type)),
-                    bytes(change_meta.domain.encode('utf-8') if change_meta.domain is not None else None),
-                    bytes(json.dumps(change_meta.to_json())),
-                )
-
-            try:
-                try:
-                    _send_to_kafka()
-                except FailedPayloadsError:
-                    # this typically an inactivity timeout - which can happen if the feed is low volume
-                    # just do a simple retry
-                    _send_to_kafka()
-            except LeaderNotAvailableError:
-                # kafka seems to be down. sleep a bit to avoid crazy amounts of error spam
-                time.sleep(15)
-                raise
-            except Exception as e:
-                _assert = soft_assert(to='@'.join(['czue', 'dimagi.com']))
-                _assert(False, u'Problem sending change to kafka {}: {} ({})'.format(
-                    change_meta.to_json(), e, type(e)
-                ))
-                raise
+            send_to_kafka(self._producer, get_topic(document_type), change_meta)
 
 
 class ChangeFeedPillow(PythonPillow):
